Move new tests to an unit test project.
