+++ test/Datadog.Trace.Tests/CallTarget/ValueTaskContinuationGeneratorTests.cs
@@ -1,12 +1,13 @@
+#if NETCOREAPP3_1_OR_GREATER
 using System;
 using System.Threading.Tasks;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.ClrProfiler.CallTarget.Handlers.Continuations;
 using Xunit;
 
-namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
+namespace Datadog.Trace.Tests.CallTarget
 {
-    public class TaskContinuationGeneratorTests
+    public class ValueTaskContinuationGeneratorTests
     {
         public static TReturn OnAsyncMethodEnd<TTarget, TReturn>(TTarget instance, TReturn returnValue, Exception exception, CallTargetState state)
         {
@@ -14,14 +15,14 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
         }
 
         [Fact]
-        public async Task SuccessTest()
+        public async ValueTask SuccessTest()
         {
-            var tcg = new TaskContinuationGenerator<TaskContinuationGeneratorTests, TaskContinuationGeneratorTests, Task>();
+            var tcg = new ValueTaskContinuationGenerator<ValueTaskContinuationGeneratorTests, ValueTaskContinuationGeneratorTests, ValueTask>();
             var cTask = tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault());
 
             await cTask;
 
-            async Task GetPreviousTask()
+            async ValueTask GetPreviousTask()
             {
                 await Task.Delay(1000).ConfigureAwait(false);
             }
@@ -33,15 +34,15 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
             Exception ex = null;
 
             // Normal
-            ex = await Assert.ThrowsAsync<CustomException>(() => GetPreviousTask());
+            ex = await Assert.ThrowsAsync<CustomException>(() => GetPreviousTask().AsTask());
             Assert.Equal("Internal Test Exception", ex.Message);
 
             // Using the continuation
-            var tcg = new TaskContinuationGenerator<TaskContinuationGeneratorTests, TaskContinuationGeneratorTests, Task>();
-            ex = await Assert.ThrowsAsync<CustomException>(() => tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault()));
+            var tcg = new ValueTaskContinuationGenerator<ValueTaskContinuationGeneratorTests, ValueTaskContinuationGeneratorTests, ValueTask>();
+            ex = await Assert.ThrowsAsync<CustomException>(() => tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault()).AsTask());
             Assert.Equal("Internal Test Exception", ex.Message);
 
-            async Task GetPreviousTask()
+            async ValueTask GetPreviousTask()
             {
                 await Task.Delay(1000).ConfigureAwait(false);
                 throw new CustomException("Internal Test Exception");
@@ -51,12 +52,12 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
         [Fact]
         public async Task SuccessGenericTest()
         {
-            var tcg = new TaskContinuationGenerator<TaskContinuationGeneratorTests, TaskContinuationGeneratorTests, Task<bool>, bool>();
+            var tcg = new ValueTaskContinuationGenerator<ValueTaskContinuationGeneratorTests, ValueTaskContinuationGeneratorTests, ValueTask<bool>, bool>();
             var cTask = tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault());
 
             await cTask;
 
-            async Task<bool> GetPreviousTask()
+            async ValueTask<bool> GetPreviousTask()
             {
                 await Task.Delay(1000).ConfigureAwait(false);
                 return true;
@@ -69,15 +70,15 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
             Exception ex = null;
 
             // Normal
-            ex = await Assert.ThrowsAsync<CustomException>(() => GetPreviousTask());
+            ex = await Assert.ThrowsAsync<CustomException>(() => GetPreviousTask().AsTask());
             Assert.Equal("Internal Test Exception", ex.Message);
 
             // Using the continuation
-            var tcg = new TaskContinuationGenerator<TaskContinuationGeneratorTests, TaskContinuationGeneratorTests, Task<bool>, bool>();
-            ex = await Assert.ThrowsAsync<CustomException>(() => tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault()));
+            var tcg = new ValueTaskContinuationGenerator<ValueTaskContinuationGeneratorTests, ValueTaskContinuationGeneratorTests, ValueTask<bool>, bool>();
+            ex = await Assert.ThrowsAsync<CustomException>(() => tcg.SetContinuation(this, GetPreviousTask(), null, CallTargetState.GetDefault()).AsTask());
             Assert.Equal("Internal Test Exception", ex.Message);
 
-            async Task<bool> GetPreviousTask()
+            async ValueTask<bool> GetPreviousTask()
             {
                 await Task.Delay(1000).ConfigureAwait(false);
                 throw new CustomException("Internal Test Exception");
@@ -93,3 +94,4 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CallTarget
         }
     }
 }
+#endif
